refactor: moved the server package to cmd, add chi routes methods
